+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v1.4.3
+
+*2024-10-31*
+
+**Changed**
+
+- Color provider for argument default values in Wikitext
+
 ## v1.4.2
 
 *2024-10-28*